# openapi.yaml
openapi: 3.0.3
info:
  title: University Services API
  version: 1.0.0
  description: 'The central API platform for university operations.

    Covers Authentication, Students, Programs & Courses, Faculty, Library, AI
    Assistant, API Key Management, and UAE Government Reporting domains.


    ## Authentication

    Most endpoints require a Bearer JWT token obtained via `POST /auth/login`.

    Government reporting endpoints (`/gov/*`) require an `X-API-Key` header instead.


    Get a JWT token: `POST /auth/login` with your email and password.

    Use JWT: Add header `Authorization: Bearer <token>` to all requests.

    Use API Key: Add header `X-API-Key: uak_<your_key>` — service role only.


    ## Test Credentials (local / seeded environment)

    | Role | Email | Password |

    |------|-------|----------|

    | Admin | admin@university.ac.ae | Admin@1234 |

    | Student | student1@university.ac.ae | Student@1234 |

    | Faculty | faculty1@university.ac.ae | Faculty@1234 |

    | Librarian | librarian@university.ac.ae | Librarian@1234 |


    ## API-Led Connectivity Architecture

    Follows MuleSoft API-led connectivity:

    - **System Layer**: Raw data access (student records, course catalog, book inventory)

    - **Process Layer**: Business logic (enrollment validation, borrowing rules, capacity
    checks)

    - **Experience Layer**: Consumer-tailored responses (student portal, faculty dashboard)

...


    ## AI Library Assistant

...
    OpenAI.

...

    '
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via OAuth2 authorization flow
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 100
      description: Number of results per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: 'Field to sort by. Prefix with - for descending. Example: sort=-created_at'
  schemas:
...
    GovMeta:
      type: object
      properties:
        total:
          type: integer
        generated_at:
          type: string
          format: date-time
        institution_code:
          type: string
          example: UAE-UNI-001
        api_version:
          type: string
          example: '1.0'
    GovStudentRecord:
      type: object
      properties:
        student_id:
          type: string
        full_name_en:
          type: string
        email:
          type: string
        enrollment_status:
          type: string
          enum: [active, inactive, graduated, suspended]
        enrollment_year:
          type: integer
        program_code:
          type: string
          nullable: true
        program_name:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        academic_level:
          type: integer
          description: UAE NQF level (1-10). Bachelor=7, Master=9, PhD=10.
          example: 7
        date_of_birth:
          type: string
          format: date
          nullable: true
    GovStudentsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/GovStudentRecord'
        meta:
          $ref: '#/components/schemas/GovMeta'
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 25
        total_items:
          type: integer
          example: 142
        total_pages:
          type: integer
          example: 6
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: The request body contains invalid fields
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  issue:
                    type: string
              example:
              - field: email
                issue: Must be a valid university email address
    StudentBase:
      type: object
      required:
      - first_name
      - last_name
      - email
      - date_of_birth
      - program_id
      - enrollment_year
      properties:
        first_name:
          type: string
          example: Nazar
        last_name:
          type: string
          example: Kamaal
        email:
          type: string
          format: email
          example: nazar.kamaal@university.ac.ae
        date_of_birth:
          type: string
          format: date
          example: '1997-05-15'
        program_id:
          type: string
          description: Reference to the Programs domain — not embedded, fetched separately
          example: prog_cs_101
        enrollment_year:
          type: integer
          example: 2026
        status:
          type: string
          enum:
          - active
          - inactive
          - graduated
          - suspended
          default: active
    Student:
      allOf:
      - $ref: '#/components/schemas/StudentBase'
      - type: object
        properties:
          id:
            type: string
            example: stu_10042
            readOnly: true
          created_at:
            type: string
            format: date-time
            readOnly: true
          updated_at:
            type: string
            format: date-time
            readOnly: true
    StudentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    Enrollment:
      type: object
      properties:
        id:
          type: string
          example: enr_80001
          readOnly: true
        course_id:
          type: string
          example: crs_5010
        student_id:
          type: string
          example: stu_10042
        enrolled_at:
          type: string
          format: date-time
          readOnly: true
        status:
          type: string
          enum:
          - active
          - dropped
          - completed
          default: active
    EnrollmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Enrollment'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    Borrowing:
      type: object
      properties:
        id:
          type: string
          example: brw_60001
          readOnly: true
        book_id:
          type: string
          example: book_301
        student_id:
          type: string
          example: stu_10042
        borrowed_at:
          type: string
          format: date-time
          readOnly: true
        due_date:
          type: string
          format: date
          example: '2026-03-25'
        returned_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          description: Set automatically by server when status changes to 'returned'
        status:
          type: string
          enum:
          - active
          - returned
          - overdue
          - lost
          example: active
    BorrowingListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Borrowing'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but insufficient permissions for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict with current resource state (e.g. duplicate email, already
        enrolled)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
- BearerAuth: []
paths:
...
  /students:
    get:
      tags:
      - Students
      summary: List all students
      description: Returns a paginated list of students. Supports filtering by program,
        year, and status.
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/SortParam'
      - name: program
        in: query
        schema:
          type: string
        example: computer_science
      - name: year
        in: query
        schema:
          type: integer
        example: 2026
      - name: status
        in: query
        schema:
          type: string
          enum:
          - active
          - inactive
          - graduated
          - suspended
      responses:
        '200':
          description: Paginated list of students
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      operationId: listStudents
    post:
      tags:
      - Students
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentBase'
      responses:
        '201':
          description: Student registered successfully
          headers:
            Location:
              description: URI of the newly created student
              schema:
                type: string
                example: /students/stu_10042
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
      operationId: createStudent
  /students/{student_id}:
    parameters:
    - name: student_id
      in: path
      required: true
      schema:
        type: string
      example: stu_10042
    get:
      tags:
      - Students
      summary: Get a specific student
      responses:
        '200':
          description: Student record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getStudent
    put:
      tags:
      - Students
      summary: Full update of student record
      description: Replaces the entire student record. All required fields must be
        included.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentBase'
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: replaceStudent
    patch:
      tags:
      - Students
      summary: Partial update of student record
      description: Updates only the provided fields. Omitted fields retain their current
        values.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                status:
                  type: string
                  enum:
                  - active
                  - inactive
                  - graduated
                  - suspended
                office_number:
                  type: string
      responses:
        '200':
          description: Student partially updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: updateStudent
    delete:
      tags:
      - Students
      summary: Remove a student
      responses:
        '204':
          description: Student deleted successfully. No body returned.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: deleteStudent
  /students/{student_id}/enrollments:
    get:
      tags:
      - Students
      summary: Get all course enrollments for a student
      description: Cross-domain relationship — links Students to Courses.
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of enrollments for this student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getStudentEnrollments
  /students/{student_id}/borrowings:
    get:
      tags:
      - Students
      summary: Get all library borrowings for a student
      description: Cross-domain relationship — links Students to Library.
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of borrowings for this student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowingListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getStudentBorrowings
...
  /courses/{course_id}/students:
    get:
      tags:
      - Courses
      summary: List all students enrolled in a course
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of students in this course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getCourseStudents
...
  /gov/students:
    get:
      tags:
      - UAE Government
      summary: UAE Gov — student roster
      description: Student list in UAE MOEHE API-first format. Requires X-API-Key with service role. JWT Bearer tokens are rejected.
      operationId: govStudents
      security:
      - ApiKeyAuth: []
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
          minimum: 1
      - name: limit
        in: query
        schema:
          type: integer
          default: 100
          minimum: 1
          maximum: 1000
      responses:
        '200':
          description: Student roster in UAE MOEHE format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovStudentsResponse'
              example:
                data:
                - student_id: abc-123
                  full_name_en: Ahmed Al Rashidi
                  email: ahmed@university.ac.ae
                  enrollment_status: active
                  enrollment_year: 2023
                  program_code: BSCS
                  program_name: Computer Science
                  department: Engineering
                  academic_level: 7
                  date_of_birth: '2000-04-15'
                meta:
                  total: 1
                  generated_at: '2026-03-15T10:00:00+00:00'
                  institution_code: UAE-UNI-001
                  api_version: '1.0'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...